+++ src/stories/mock/Flexible_CallToAction.js
@@ -1,0 +1,15 @@
+export const mock = {
+    id: "16983",
+    typeHandle: "callToAction",
+    callToAction: [
+        {
+            id: "16984",
+            titleCta: "Call to action",
+            summary: "a call to action",
+            icon: "svg-call-to-action-find",
+            buttonText: "click me",
+            buttonLink: "http://google.com",
+            backgroundColor: "false",
+        },
+    ],
+}