+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [1.49.1](https://github.com/UCLALibrary/ucla-library-website-components/compare/v1.49.0...v1.49.1) (2022-08-19)
+
+
+### Bug Fixes
+
+* divider margins ([#128](https://github.com/UCLALibrary/ucla-library-website-components/issues/128)) ([e04491c](https://github.com/UCLALibrary/ucla-library-website-components/commit/e04491c1dcad3ea14763d0842d4d7bca0e301e06))
+
+# [1.49.0](https://github.com/UCLALibrary/ucla-library-website-components/compare/v1.48.4...v1.49.0) (2022-08-19)
+
+
+### Features
+
+* Refactor FlexibleBlocks ([#112](https://github.com/UCLALibrary/ucla-library-website-components/issues/112)) ([ac3b6c1](https://github.com/UCLALibrary/ucla-library-website-components/commit/ac3b6c179384a8777ece4e0a57846b531bb39f40))
+
 ## [1.48.4](https://github.com/UCLALibrary/ucla-library-website-components/compare/v1.48.3...v1.48.4) (2022-08-18)